+++ src/lib.rs
@@ -29,7 +29,9 @@ mod auth;
 #[cfg(feature = "client_api")]
 pub use crate::auth::AuthError;
 #[cfg(feature = "client_api")]
-pub use crate::gitlab::{AsyncGitlab, Gitlab, GitlabBuilder, GitlabError, RestError};
+pub use crate::gitlab::{
+    AsyncGitlab, Gitlab, GitlabBuilder, GitlabError, ImpersonationClient, RestError,
+};
 pub use crate::types::*;
 
 #[cfg(test)]